04-MINI-PROJECT/DRF-VUE2-MOVIEAPP/backend/requirements.txt
asgiref==3.5.2
black==22.10.0
click==8.1.3
colorama==0.4.5
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.12.2
mypy-extensions==0.4.3
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
pytz==2022.4
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.5
